refactor: AI 챗봇 구조 개선 및 Redis Pub/Sub 로컬 프로필 분리
