+++ src/main/java/org/example/oddventure/domain/ai/controller/ChatbotController.java
@@ -1,15 +1,13 @@
 package org.example.oddventure.domain.ai.controller;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.example.oddventure.common.dto.response.ApiResponse;
 import org.example.oddventure.domain.ai.agent.AgentExecutor;
 import org.example.oddventure.domain.ai.agent.AgentRunnerService;
-import org.example.oddventure.domain.ai.pubsub.RedisPublisher;
-import org.example.oddventure.domain.ai.service.LoadTestChatbotService;
+import org.example.oddventure.domain.ai.dto.UserMessage;
 import org.example.oddventure.domain.auth.dto.AuthUser;
-import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,38 +17,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/chat")
-@Profile("local")
 public class ChatbotController {
 
     private final AgentRunnerService agentRunnerService;
-    private final LoadTestChatbotService loadTestChatbotService;
-    private final RedisPublisher redisPublisher;
+    private final SimpMessagingTemplate messagingTemplate;
 
-    // 로컬 테스트용
     @PostMapping
     public ResponseEntity<ApiResponse<AgentExecutor.State>> reply(
             @AuthenticationPrincipal AuthUser user,
             @RequestBody UserMessage userMessage
     ) throws Exception {
-        // 테스트용 Redis 직접 발행
-        String channel = "chat:" + user.id() + ":input";
-        redisPublisher.publish(channel, userMessage);
-
         AgentExecutor.State reply = agentRunnerService.execute(user.id(), userMessage.message()).orElse(null);
+        messagingTemplate.convertAndSend("/topic/chat/" + user.id(), reply);
+
         return ApiResponse.success(reply, "AI 응답 테스트가 정상적으로 완료되었습니다.");
-    }
-
-    @PostMapping("/loadtest")
-    public ResponseEntity<ApiResponse<String>> loadTest(
-            @RequestBody UserMessage userMessage
-    ) {
-        String reply = loadTestChatbotService.reply(userMessage.message());
-        return ApiResponse.success(reply, "부하 테스트용 응답입니다.");
-    }
-
-    public record UserMessage(
-            @NotBlank(message = "메시지를 입력해주세요!")
-            String message
-    ) {
     }
 }
