+++ src/main/java/org/example/oddventure/domain/ai/dto/UserMessage.java
@@ -1,0 +1,9 @@
+package org.example.oddventure.domain.ai.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record UserMessage(
+        @NotBlank(message = "메시지를 입력해주세요!")
+        String message
+) {
+}
