+++ src/main/java/org/example/oddventure/domain/ai/pubsub/ChatMessageOutputSubscriber.java
@@ -2,11 +2,13 @@ package org.example.oddventure.domain.ai.pubsub;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
+@Profile("local")
 @Slf4j
 @Service
 @RequiredArgsConstructor
