+++ src/main/java/org/example/oddventure/domain/ai/pubsub/ChatMessageSubscriber.java
@@ -3,10 +3,12 @@ package org.example.oddventure.domain.ai.pubsub;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.oddventure.domain.ai.service.ChatbotService;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.stereotype.Service;
 
+@Profile("local")
 @Slf4j
 @Service
 @RequiredArgsConstructor
