+++ src/main/java/org/example/oddventure/domain/ai/pubsub/RedisPublisher.java
@@ -2,14 +2,11 @@ package org.example.oddventure.domain.ai.pubsub;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
-/**
- * Redis Pub/Sub 발행자
- * <p>
- * 서버 내부에서 발생한 이벤트를 Redis 채널로 발행
- */
+@Profile("local")
 @Slf4j
 @Component
 @RequiredArgsConstructor
@@ -17,7 +14,6 @@ public class RedisPublisher {
 
     private final RedisTemplate<String, Object> redisTemplate;
 
-    // 지정한 채널로 메시지를 발행
     public void publish(String channel, Object message) {
         redisTemplate.convertAndSend(channel, message);
         log.info("[Redis 발행] channel={}, message={}", channel, message);
